move mock into test class
