+++ src/test/java/sk/uniba/fmph/dcs/game_phase_controller/PlaceFiguresStateTest.java
@@ -8,6 +8,51 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.junit.Assert.*;
+class FigureLocationMock implements InterfaceFigureLocation {
+
+    ActionResult expectedActionResult;
+    HasAction expectedHasAction;
+    Boolean expectedBoolean;
+
+    @Override
+    public boolean placeFigures(PlayerOrder player, int figureCount) {
+
+        return expectedBoolean;
+    }
+
+    @Override
+    public HasAction tryToPlaceFigures(PlayerOrder player, int count) {
+
+        return expectedHasAction;
+    }
+
+    @Override
+    public ActionResult makeAction(PlayerOrder player, Effect[] inputResources, Effect[] outputResources) {
+
+        return expectedActionResult;
+    }
+
+    @Override
+    public boolean skipAction(PlayerOrder player) {
+
+        return expectedBoolean;
+    }
+
+    @Override
+    public HasAction tryToMakeAction(PlayerOrder player) {
+
+        return expectedHasAction;
+
+    }
+
+    @Override
+    public boolean newTurn() {
+
+        return expectedBoolean;
+
+    }
+}
+
 public class PlaceFiguresStateTest {
 
     private PlaceFiguresState placeFiguresState;
